Handle non-zero exit status in python
Return errors back to Python in exitstatus and convert to an exception there.
This uses something very like a decorator to wrap the Fortran function in
a Python function during module load (this could be automated for all
Fortran functions in a similar way to how we add documentation already). This
avoids messages like:

    capi_return is NULL
    Call-back cb_pystop_in_PlmBar__user__routines failed

and allows more controll over the nature of the exception (although this
is not implemented). At present the traceback ends up pointing at the
wrapper function but it may be possible to strip this out so the innermost
frame points at the call into the shtools function without making the internals
so visible.